Removed a code optimization that no longer helps
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@481 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/cpu/CPU.java
@@ -508,7 +508,7 @@ public class CPU extends Module_base {
     static public /*Bitu*/int Segs_FSphys;
     static public /*Bitu*/int Segs_GSphys;
 
-    static public long seg_value(int index) {
+    static public int seg_value(int index) {
         switch (index) {
             case CPU_Regs.es:
                 return Segs_ESval;
